refactor(search): code-split searchbar
