+++ components/Searchbar/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Searchbar';
+export * from './Searchbar';
